feat: unit test rework

+++ src/BurgerBackend.Api/Contracts/Models/Scoring.cs
@@ -4,5 +4,5 @@ public class Scoring
 {
     public string ScoringName { get; set; } = string.Empty;
 
-    public decimal ScoreValue { get; set; }
+    public double ScoreValue { get; set; }
 }
